refactor(ui): split useHeaderActions into thematic sibling hooks
useHeaderActions.ts: 546 → 401 lines. Three self-contained concerns
moved to sibling hooks:

  - useFullscreen — Tauri + browser fullscreen state + toggle
  - useFilenameEditor — inline filename editor state (ref, edit/blur/keydown)
  - useHeaderKeyboardShortcuts — keydown effect wiring shortcut ids to handlers

The orchestrator's public return shape is unchanged (Header.tsx
consumer untouched).
